requirements.txt
Pyrebase4==4.5.0
dash==1.21.0
Flask-Caching==1.9.0
dash-bootstrap-components==0.11.1
dash-daq==0.5.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-extensions==0.0.44
